Origin/feature/view

+++ Sindie.ApiService/Sindie.ApiService.Core/Entities/BodyTemplatePart.cs
@@ -92,7 +92,7 @@ namespace Sindie.ApiService.Core.Entities
 			BodyTemplate bodyTemplate = default,
 			BodyPartType bodyPartType = default,
 			string name = default,
-			int damageModifier = default,
+			double damageModifier = default,
 			int hitPenalty = default,
 			int minToHit = default,
 			int maxToHit = default,
